requirements.txt
asgiref==3.5.2
Django==3.2.8
django-appconf==1.0.5
django-compressor==4.1
django-crispy-forms==1.14.0
django-social-share==2.3.0
django-taggit==3.1.0
django-tinymce==3.4.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pytz==2022.6
rcssmin==1.1.0
rjsmin==1.2.0
sqlparse==0.4.3
